Deduplicate media type fixture
